Bugfix/delete user by

+++ src/main/java/org/example/schedule/user/controller/UserController.java
@@ -1,11 +1,13 @@
 package org.example.schedule.user.controller;
 
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpSession;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
-import org.example.schedule.common.filter.LoginFilter;
 import org.example.schedule.common.session.SessionConst;
 import org.example.schedule.user.dto.*;
 import org.example.schedule.user.service.UserService;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -27,13 +29,21 @@ public class UserController {
         return ResponseEntity.ok(userService.findUsers(username));
     }
 
-    @PutMapping("/users/{userId}")
+    @PutMapping("/users/me")
     public ResponseEntity<UserUpdateResponse> updateUser(@SessionAttribute(name = SessionConst.LOGIN_USER) long userId, @RequestBody UserUpdateRequest request) {
         return ResponseEntity.ok(userService.update(userId, request));
     }
 
-    @DeleteMapping("/users/{userId}")
-    public void deleteUser(@PathVariable long userId) {
+    @DeleteMapping("/users/me")
+    public ResponseEntity<Void> deleteUser(HttpServletRequest request) {
+        HttpSession session = request.getSession(false);
+        if (session == null || session.getAttribute(SessionConst.LOGIN_USER) == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+
+        Long userId = (Long) session.getAttribute(SessionConst.LOGIN_USER);
         userService.delete(userId);
+        session.invalidate();
+        return ResponseEntity.ok().build();
     }
 }
